requirements.txt
telethon>=1.36.0
python-telegram-bot>=21.6
SQLAlchemy>=2.0.36
pydantic>=2.9.2
openai>=1.54.0
httpx>=0.27.2
beautifulsoup4>=4.12.3
tenacity>=9.0.0
tqdm>=4.66.5
python-dateutil>=2.9.0.post0
